chore: 예약 dto 테마/시간 아이디 기반으로 수정
